+++ all-is-cubes/src/raytracer/updating.rs
@@ -128,6 +128,7 @@ struct SrtTodo {
 }
 
 /// [`Listener`] adapter for [`SpaceRendererTodo`].
+#[derive(Clone, Debug)]
 struct TodoListener(Weak<Mutex<SrtTodo>>);
 
 impl Listener<SpaceChange> for TodoListener {